add 'revision' attribute to git_commit
This is a breaking change in order to support different ways to specify a single commit, e.g. by using 'HEAD'. The old 'sha1' is now read-only and populated by this provider after the given revision was resolved.

In order to migrate to the new version simply change 'sha1' to 'revision' in the declaration of your data source. The 'sha1' output stays the same, therefore no change is required in dependent resources.

Relates to #28